Expand optimistic update to order_article.
What is missing now, is the computation of the new group_order_article's
result from the updated order_article's `units_to_order`. This will be
updated after the group_order_article update response (not optimistic).
It does affect the price shown, so it would be nice to update that
optimistically too.